HBX-1601: Move class 'org.hibernate.cfg.binder.BinderUtils' to 'org.hibernate.tool.internal.reveng.BinderUtils'
